third_party/WebKit/LayoutTests/fast/events/dragevent-constructor.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the constructor of the standard drag event");

var testEvent;
shouldNotThrow("testEvent = new DragEvent('dragstart')");
shouldBe("testEvent.__proto__", "DragEvent.prototype");
shouldBe("testEvent.dataTransfer", "null");

shouldNotThrow("testEvent = new DragEvent('dragstart', { dataTransfer:null })");
shouldBe("testEvent.__proto__", "DragEvent.prototype");
shouldBe("testEvent.dataTransfer", "null");
</script>
