third_party/blink/web_tests/fast/events/dragevent-constructor-expected.txt

Tests the constructor of the standard drag event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testEvent = new DragEvent('dragstart') did not throw exception.
PASS testEvent.__proto__ is DragEvent.prototype
PASS testEvent.dataTransfer is null
PASS testEvent = new DragEvent('dragstart', { dataTransfer:null }) did not throw exception.
PASS testEvent.__proto__ is DragEvent.prototype
PASS testEvent.dataTransfer is null
PASS successfullyParsed is true

TEST COMPLETE
